Commit 2
- Create NewsService class
- Put Array list in NewsService and make an instance in the main activity.
(not 100% sure yet if this is what is being asked in the assignment)
